Update SEDML to latest output from copasi.sedml
In many cases, the only change is the Copasi version number.  The other common changes are to Copasi's numbering of ranges (they now start at 'range0' instead of 'range1'), and its labeling of various elements, hopefully to ensure uniqueness (which had been a problem).

In some (few) cases, the model source was correct in the original SEDML file (instead of just being 'model.xml').  In those cases, the correct model source has been preserved.  However, in several cases, the id of that model was also changed, without a corresponding change to references to that model.  In those cases, the 'source' was changed, and the 'id' was not.

In some other cases, the new SEDML file produced was substantially different from the previous version; more than just labels, from small differences like 'timeEnd' of a simulation to larger changes like 'several dataGenerators' are present in one but not the other'.  In ten cases, the original model was either obviously better (more complete) or simply different.  In those cases, the originals were preserved.  In 13 cases, the new file was obviously better, often simply by dint of having data generators and output at all, instead of having none.  In those cases, the new ones are here checked in.