Remove omex files that contain no extra information beyond what's in the main directory.
All of these files can either be re-generated from the copasi files, or are out of date from the copasi files, and would have needed to be updated anyway.  (Most of the more recent ones had completely empty sedml files in them for example.)

There is exactly one omex file (585) that was not generated from copasi, and also contains files not in the main directory.  We'll need to figure out how to curate this one specially.