Fix YAML strings.
My python choked trying to parse the old/new strings that were set up with carriage returns in them.  Since the returns are not actually needed to uniquely find the problem in question, this converts them to match all the other strings in the file.

(Also, adds a return at the end.)