Validate and fix SBML files.
The validation takes some time, so it's put in as an optional argument to fix-entries.  It was then added as an argument to the github actions, but what's still missing is a way to then check the results to make sure no new invalid SBML files have been created.

Also, updateSBML.py was added which fixes the validation errors found.  The vast majority of problems were unit errors, which was 'solved' by converting the files to SBML L1v4, where unit validations were demoted to warnings instead of errors.  The other errors fixed were:

* Biomodels 344, MODEL1005280000_moderatestress: turn a species that both had an assignment rule and appeared in reactions to be a boundary species.
* Biomodels 439 and 479:  change the substanceUnits of species to nanomole instead of nanomolar.
* Biomodels 631:  Remove invalid SBO terms.

Also:
* moved a file rename from fix_manual_corrections.yml to fix_filenames.py so that it would hit before the construction of the SBML list.
* Removed bits of the manual_corrections that I made directly to manual-fixes/
* Made one more layout fix I had missed last time.
* Re-enable fix_namespaces_in_sedml_doc which I had accidentally broken earlier.
* Updated manual-fixes.txt

The commit actually changes a lot more than that, since I still can't run the convert_entry function, but anyway.