Fix for #83: fix identifier.org references.
In particular:
* Get valid pubmed references.
* Make sure all the identifiers.org root namespaces are real.

Many 'blank' references had to be removed by hand (the manual-fixes directory), as well as URLs with accidental spaces in them.

There are still a few namespaces that are left but that aren't real:  omim, psimi, and bind.  The first two are real places that need identifiers.org to recognize them.  The last I don't yet know what it is.