Attempt to fix more character encoding problems.
Many of these change make no sense to me--I change one part of the file, and suddenly a different part of the file is fixed.  But hey, I guess character encoding is hard!

One particularly odd bit is that I had to change the author lists in 1011 and 1012 to not have any special characters, but the author lists in 1013 and 1014 were fine, despite having many of the same characters.  And, what's more, copying the author lists from 1013 to 1011 still was problematic for those files!