Add ModelChange elements to all models without any.
Fix for issue #101:  for every model with no ModelChange, add a set of ModelChanges that change all changeable elements in the model.  Use the biosimulators-utils pre-exisiting function for this.

Also remove 'comp' from model 613.  It's not actually used, and confuses the poor utils function.

Also, bump all SedML versions to l1v4.  This allows the modelChange elements to actually have ids and names (as helpfully provided by the utils function), and will be useful later when we add some extra plot information.