First pass at including plot information from Copasi.
We definitely need to know if a given plot is scatter or a line.  This is a first pass at creating the l1v4 constructs needed for this.  Still missing is a way to extract that information from Copasi, and a way to match the plot in the SED-ML with the plot object in Copasi.