Finish fix for #82:  get plot info from Copasi.
Thanks to Frank for how to get this, we now can read in the plot information from Copasi and use it to set a style for the curves.  Most styles are 'a line with thickness 1' but some are different.  The most important thing to check is line type and marker type, because if there's no line, the line type is set to None.

Also remove a backup file that got committed accidentally, and other changes to fix_manual_corrections.yml that came about because of changing the SED-ML to L1v4, which changed 'numberOfPoints' to 'numberOfSteps', and made a couple of fixes impossible (for me) to encode in YML because the indenting of the fix matched the indenting of the YML itself.  Those fixes moved to recreate_sedml_from_copasi.py instead.