Adjust scripts to account for updated Copasi.
Copasi now produces better SED-ML, so some of the fixes we made need to be adjusted.  In particular:
* Nested parameter scans are now produced by default in tasks.  Unfortunately, the dataGenerators point to the *innermost* repeated task, not the outermost, so that part still needs to be fixed.  (This bug filed with @frankbergman
* Uniform time courses now export the intended output time and not the time you get by multiplying the number of intervals by the interval length.  We still often want to change them (usually because the number of intervals is weird) but the replaced text is now closer to the final text.

Because we rename some datagenerators because of the first point, some of the 'revised output names' had to be similarly adjusted.

Something also seems to have happened with algorithm parameter names--an earlier update of libsedml had been adding in useful names when setting Kisao terms, but this seems to be turned off again.  It's not essential, but hopefully further updates will turn this back on again.