Update errors and error messages.
If the SFBC converter can't be found, this now raises an error.  When problems arise, we now report the error instead of claiming that the SBML is probably invalid.

(We may want to drop that claim entirely and just report the actual error, particularly since in my own experience both on my laptop and with github actions, the issue is never SBML validity.)