Fix SBML file copying issues.
If we renamed SBML files in fix_manual_corrections, that's run after the create-SEDML step, so the SED-ML ends up with the wrong filenames.  So we have to create a separate SBML-renaming scheme that runs before fix_manual_corrections.

Also, while looking at renamed SBML files, saw that there was an errant 'fig14..xml' in the list, so removed that.  Tried to remove it with the renaming scheme, but that caused other problems.