Update for latest release of Copasi
The latest Copasi changes how it outputs SED-ML again, for the better.  This means we no longer need to make as many corrections, but also means that some of the corrections we do still need to make have different targets.

The biggest change is that Copasi now can output plot information (styles), so we no longer need to create our own.

In addition, Copasi now treats parameter scans fully correctly, and makes data generators that reference the 'outermost' repeated task.

Other than that, there are a few changes in the names of variables here and there.