Fix for new parallelization scheme.
The current scheme runs models 0-40 in one thread, models 1-41 in another, etc.  I have no idea if you can use math in this way for the github workflows, so this is an experiment.

The next problem to solve will be only committing the 'final/' models that were actually changed, but let's see if the first thing works first.

Reduce number of jobs running at once while testing.