Fix generated sedml from 734
The Copasi file for Biomodel 734 seems to have been set up as a parameter estimator, which Copasi does not yet actually output to SED-ML (and the external data is missing, anyway).  But the simulation is good, so delete all the incorrect data generators, etc., and replace with the autogenerated ones.