Speed up checking pubmed ids.
The last two biomodels in particualar have thousands of them, but only ~1000 *unique* ones, so just go through the ones that are new.

This could be sped up even more if we only *ever* checked them once, and stored the good ones in a local db, but this works for now.