Workaround for copasi bug in generating SED-ML
It turns out that if you ask Copasi to translate a model with a function definition with a given id (in this case, "M1" from model 757), it will then forbit the use of "M1" from any other model you ask it to translate.  This caused it to erroneously rename species "M1" from models 769 and 770 to "M1_0".

Moving 757 to the end of the list is a workaround for this problem for now, but I've filed it with the Copasi team, because it might bite us later if we end up with different sublists of models to translate.