Remove always-NaN observable.
For some reason, the parameter 'Curve_Va_Xa' was set up to have an assignment rule Xa_Va/Metabolite_13.  In theory, Xa_Va should be able to change in time, but Metabolite_13 is the initial value of Xa_Va, which is zero, so the only possible values for this 'curve' were NaN or INF.  As such, I have removed the parameter entirely.